reorganise the wallet command and tooling
