+++ cardano-cli/src/wallet/error.rs
@@ -1,0 +1,10 @@
+use cardano::hdwallet;
+
+/// wallet errors
+pub enum Error {
+    CannotRetrievePrivateKeyInvalidPassword,
+    CannotRetrievePrivateKey(hdwallet::Error),
+}
+impl From<hdwallet::Error> for Error {
+    fn from(e: hdwallet::Error) -> Self { Error::CannotRetrievePrivateKey(e) }
+}
