+++ cardano-cli/src/wallet/result.rs
@@ -1,0 +1,4 @@
+use super::error::Error;
+
+/// alias for Wallet operation result
+pub type Result<T> = ::std::result::Result<T, Error>;
